requirements.txt
apache-beam-with-gcp==2.41.0
fsspec==2022.8.2
gcsfs==2022.8.2
google-api-python-client==2.58.0
google-cloud-storage==2.5.0
pandas==1.4.4
python-dotenv==0.20.0
pandas-gbq==0.17.8
spotipy==2.20.0
